+++ webapp/tests/unit/user/user-extension-schema-defaults.test.js
@@ -3,8 +3,8 @@
  * @tagline         applyExtensionSchemaDefaults for GET /api/1/user
  * @description     Merges plugin/site extendSchema defaults into API user payloads
  * @file            webapp/tests/unit/user/user-extension-schema-defaults.test.js
- * @version         1.6.43
- * @release         2026-04-22
+ * @version         1.6.44
+ * @release         2026-04-23
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/